RetailTalk/backend/requirements.txt
# SmartShop Backend - Python Dependencies

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# Database
supabase==2.7.2
psycopg[binary]==3.3.2

# ML / AI
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0

# Auth
PyJWT==2.9.0
bcrypt==4.2.0
python-multipart==0.0.9

# CORS & other
pydantic[email]==2.9.0
